+++ CMakeLists.txt
@@ -205,7 +205,7 @@ set(SOURCE_FILES
         src/FileList/FileInfoLoader.cc
         src/FileList/FileListHandler.cc
         src/FileList/FitsHduList.cc
-        src/Frame.cc
+        src/Frame/Frame.cc
         src/GrpcServer/CartaGrpcService.cc
         src/ImageData/CartaFitsImage.cc
         src/ImageData/CartaHdf5Image.cc
