+++ test/TestFitsImage.cc
@@ -7,7 +7,7 @@
 #include <gtest/gtest.h>
 
 #include "ImageData/FileLoader.h"
-#include "src/Frame.h"
+#include "src/Frame/Frame.h"
 
 #include "CommonTestUtilities.h"
