+++ test/TestHdf5Attributes.cc
@@ -8,7 +8,7 @@
 #include <gtest/gtest.h>
 
 #include "ImageData/Hdf5Attributes.h"
-#include "src/Frame.h"
+#include "src/Frame/Frame.h"
 
 #include "CommonTestUtilities.h"
