+++ test/TestHdf5Image.cc
@@ -8,7 +8,7 @@
 #include <gtest/gtest.h>
 
 #include "ImageData/FileLoader.h"
-#include "src/Frame.h"
+#include "src/Frame/Frame.h"
 
 #include "CommonTestUtilities.h"
