+++ test/TestSpatialProfiles.cc
@@ -10,7 +10,7 @@
 #include "CommonTestUtilities.h"
 #include "ImageData/FileLoader.h"
 #include "Util/Message.h"
-#include "src/Frame.h"
+#include "src/Frame/Frame.h"
 
 using namespace carta;
